have media.codec hold dlopen() codecs forever
use RTLD_NODELETE so that the any codecs are kept forever.
This dramatically shortens the time for a subsequent dlopen()
for the same library. It turns the dlclose() of that library into a NOP.

This eliminates most dlopen/dlclose cpu time as media.codec processes
the same content type without restructuring the registration and de-registration
logic that is in place.

Bug: 36727951
Change-Id: If92bc2898f8c73f89f164524daffab81da17002e
Test: getting continued video/sound across multiple invocations