Camera: Correct ae regions request for metering area
The size of ae regions may be zero by rounding error.
But it should be at least 1 if metering area size is not zero.

This problem causes CTS fail for
android.hardware.cts.CameraTest#testMeteringAreas on
the device which has the active array width less than 3556
if the preview size is 1920x1080.

Bug: 63795820
Test: ran Camera CTS
Change-Id: Ie3995e40aedba193393df8fd05d95f740a25ef3d