Camera: Device 3: Ignore buffer limits when clearing request queue
When abortCaptures is called, the request queue is cleared; as part of that,
any requests with input buffers trigger the removal and return of one input
buffer from the input stream.

However, if the HAL currently is processing some number of reprocess
requests, the stream's max buffer limit may have been reached, in
which case getInputBuffer will block until the HAL returns an input
buffer.  This stalls flushing (and calling of capture_request) for
some time, and seems to cause problems for the HAL during the flush.

So instead, don't respect the HAL's max buffer limit when all we're
doing is throwing work away.

Test: CTS, manual testing
Bug: 62420820
Change-Id: I72d8cdaf67fb3cc6876a03cee9e0021d95cecdfe