Have the Surface class track the buffer age.
This change corresponds to a change between the Surface and
BufferQueueProducer classes.

Have the Surface class track the buffer age, so that Surface::query()
can return the buffer age without having to use a binder call to
BufferQueueProducer::query().  The idea is for
BufferQueueProducer::dequeueBuffer() to return the value, which the
Surface class will cache for later use by Surface::query().

Bug: b/27903668
Test: Use systrace to no ensure query binder call after dequeueBuffer.
Change-Id: Iee7e7a3f92dcc88db70f460186a63450a6f18e31