audiopolicy: do not check for mode_in_communication for voip flags
Some VoIP clients may choose to open the streams as soon as the phone
state is set to MODE_RINGTONE to allow lower start latency when the call
is connected. This patch removes the unnecessary check conditions for
MODE_IN_COMMUNICATION since the stream type is sufficient indicator for
setting the flags for VoIP calls.

Bug: 62782025
Test: manual voip mo and mt calls tests
Change-Id: Idc7e1b6f33921580c0937dee71b2b2105cbd31eb
Signed-off-by: David Lin <dtwlin@google.com>