Update patch set 1
Patch Set 1:

It can't be app permission, because WifiDisplay is all integrated and controlled by Android. This is not Netflix or anything like that. WifiDisplay clones your phone screen and audio to a TV. If anything it might be audio HAL, but I don't really believe it is. Someone should test it on newer HALs if possible, to confirm.

Patch-set: 1