Codec2: make C2Status into an enumeration
This is to distinguish it from android::status_t and to avoid implicit
conversions.

As part of this change move status_t return values to C2Status, rename
the old C2Error to the new C2Status name to better reflect its purpose.

Kept block::error() method names as those are to be used to
check for any errors after an implicit bool cast, e.g. if (view.error)...

Bug: 64121714
Test: unittest
Change-Id: I09d8e28cf63cdd9d0ac9fd29a8da502b89c52f81