add test for AudioTrack constructor
This test contructs a number of AudioTracks from arguments listed in an input file
and outputs dumps of the constructed AudioTracks in an output file.
The output file can be compared to previous iterations to spot regressions.

An reference input file and the corresponding output file is added for
further comparison.

Test: run this test
Change-Id: I88e458a194a18c9ec919e1e903d49686632c18bf