add test for AudioRecord constructor
This test contructs a number of AudioRecords from arguments
listed in an input file and outputs dumps of the constructed
AudioRecords to an output file.
The output file can be compared to previous iterations to spot regressions.
An reference input file and the corresponding output file are added for
further comparison.

Test: run this test
Change-Id: I0d1b596ca91c3c58f159887fbd48aa3f04ae38c2