AudioPolicy: enable phone apps to mute VOICE_CALL stream
This change enables an app that has the MODIFY_PHONE_STATE
permission to mute the VOICE_CALL stream.
The change adds specific handling in  AudioPolicyManager
for the case in which volume index is 0 and stream type is VOICE call.
The permission check is done in the AudioService Java code for the
calling process.

Bug: 71017199
Fixes: 71017199
Test: Tested manually
Change-Id: Ibbf3b6948a4384a963d00347ebd04ccf6d56ebd5