[2/2] libstagefright: Add support for NV21 color format
This adds the native color format 'OMX_SEC_COLOR_FormatNV21Linear' of
exynos5 which is NV21 and needed for the camera. You can turn it on
with:

BOARD_USE_SAMSUNG_COLORFORMAT_NV21 := true

It requires

const char CameraParameters::PIXEL_FORMAT_YUV420SP_NV21[] = "nv21";

to be specified in libcamera_parameters.

Change-Id: Iccf287244cc6376d53975daf1793bb7861b859eb