soundtrigger: Add support for HAL V2.1
Use HAL V2.1 methods if this version is deployed on the device.
Initially the middleware layer requests for HAL V2.0, but if
the interface can be cast to V2.1, then new methods are used.

Made changes to avoid C-style type coercion as it was leading
to erroneous results when using V2.1 and V2.0 data structures
simultaneously.

Bug: 68823037
Test: use "ok google" and music IQ with HAL V2.0 and V2.1
Change-Id: I34924afe3cd4fb45a75fa34888e6e5c72727dbce