Codec2: fix up C2Param
Add an internal _C2ParamInspector to be able to inspect fields of
C2ParamField and C2FieldId, and to parse C2Param out of blob.

Fix C2Param::updateFrom to allow copying a param that is smaller
than the current param (e.g. for flex params) as that is always
safe.

Allow enum helper macros to be 'preinited' by C2Config.h to allow
any include order.

Bug: 64121714
Change-Id: I8be8fa974194bfe397f90a334e786f94fe20862c