MediaControlView2: Add Support for Ad
This CL adds additional UX for showing the remaining time and the
unskippable time. Also disable NEXT button for the first 5 seconds
and disable SEEK throughout the whole duration of the advertisement.

Also, this CL adds a listener to TitleBar so that when the layout of
VideoView2 is changed the layouts of the TitleBar is also changed to
avoid overlapping text.

Bug: 73136129
Test: run VideoViewTest.apk
Change-Id: Ic91b3f285343f4b6bbca2900ce499fb1f0d54bf6