Audio V4: *CNT and *MAX are no longer part of the enum
audio.h used to have *_CNT and *_MAX values in its enum
mostly for enum iteration.

This was not compatible with Treble as it denied any enum
inheritance.

*_CNT and *_MAX are now implemented with macro that can
be updated and are not part of the API.

Bug: 38184704
Test: Compile
Change-Id: I3921fdf3435ba114ca77a4595ba8fc97a5e0438d
Merged-In: I3921fdf3435ba114ca77a4595ba8fc97a5e0438d
Cherry-picked from master
Signed-off-by: Kevin Rocard <krocard@google.com>