heif: use width and height for decoder -- DO NOT MERGE
The display dimensions from MPEG4Extractor is applied on top
of width and height for scaling, it's not for cropping. For
bitmap uses we have to use width and height, the frame from
media server is not scaled.

bug: 73172046
Test: test app in bug; open folders with heif files in Photos.
Change-Id: I3d7f1492d08d48876836ccc05b6eb4de0d0c0f9a
(cherry picked from commit 1a6c11685d3868d0ebe8543842d213c4ff7e9009)