MediaSession2: Avoid binding from system service if able
MediaController2 had bounded to session service whenever it's created.
It was to increase binding counter for a session service to keep it
running while any controller exists, but it doesn't make sense for the
system service which creates controllers only to monitor changes in a
session.

Avoid binding from system service if able.

Bug: 73872399
Test: Run all MediaComponents tests
Change-Id: I16c81d67771e3716b646c8a48df5411bb4cf0527