CameraService: Handle ActivityManager death for UidPolicy
Since the camera service is not killed when the system service dies,
it needs to handle detecting the death and re-registering for
UidPolicy updates when the system service has restarted.

When no activity manager is available, don't limit access by UID
status.

Use calls from the CameraServiceProxy service in system service to
trigger re-registration of UidPolicy in case of system service death.

Test: adb shell stop; adb shell start; verify camera service allows
    camera use. No camera CTS regressions.
Bug: 74230547

Change-Id: I8ff6b1e88117cc99b9b85e4069a947ff99236e1d