Update patch set 1
Patch Set 1:

> (1 comment)

There are two conscructors setting this parameter to true. The other one is CameraSourceTimelapse. By setting this parameter in the init function, the change to achieve the desired behaviour remains minimal.

A "totally clean" fix would require to also remove the default parameter from the Constructors. But I don't think we want to add that much changes in code coming from upstream.

Patch-set: 1