OMXNodeInstance: Allow skipping CVE-2017-13154
This commit breaks the audio and/or video playback on some devices.

Wrap into #ifndef/#endif statement to allow skipping this patch for
specific devices. The ifndef was added on Android 6 (Change 198403)
directly into the CVE-2017-13154 patch.

Add the following directive into the BoardConfig.mk or
BoardConfigCommon.mk file of the device repo to skip this change:
  BOARD_SKIP_CVE_2017_13154 := true

cm-11.0: use a Boardflag instead C/C++ flag

Change-Id: I9796dc5b1389668c33e7aebbe703baa1244192c8