OMXNodeInstance: Add ifndef for CVE-2017-13154 patch
The ifndef was added on Android 6 (Change 198403) directly on
CVE-2017-13154 patch.

This commit breaks the audio and/or video playback on some devices.
Wrap into #ifndef/#endif statement to allow skipping this patch for
specific devices by adding the following directive into the
BoardConfig.mk or BoardConfigCommon.mk file of the device repo:
  TARGET_RELEASE_CPPFLAGS += -DSKIP_CVE_2017_13154

Change-Id: I9796dc5b1389668c33e7aebbe703baa1244192c8