Mark libmedia_omx and libstagefright_foundation as double_loadable
These two are explicitly marked as double_loadable since they are one of the
dependencies of the LLNDK library libnativewindow (libmediandk)
and at the same time the lib themselves are marked as VNDK. Such libs can be
double loaded inside a vendor process.

Note: even without this change, the library is already capable of being
double loaded due to the dependency graph around it. This change is to
make it explicit so that double loading of a library is carefully
tracked and signed-off by the owner of the lib.

Bug: 77155589
Test: m -j
Change-Id: I38e3ff4a76930a99a04f3dbc53a2dcfea18de8f3