Update patch set 4
Patch Set 4:

On your personal notes: You need to stop acting as if I'm out to get you, out to destroy the project, out to kill legacy devices, etc. I don't take offense easily, but that's an accusation that I will take offense to. It's also hilarious to me that a maintainer would equate a device being official with being supported, or a maintainer would think that having an official device is required in order to actively maintain a device or contribute to the platform.

On the patch: The goal of me commenting on this patch is to get an idea of where we stand on if this is a hack that we would accept on shipping devices. Why would I ask about that? Because we have standards for device support, and I think this is a blatant abuse of the lenient wording that exists. I'm certainly not blocking this patch from being merged (nor do I have the gerrit power to do so), but I'm stating my concern (which I have every right to do), and explaining why I have such a concern. So, If the general consensus is that we wouldn't accept this hack on a shipping device, then there's no reason for us to merge this patch in the first place.

As a side note, I will not approve this being used for any of my legacy hal1 devices, and I think Kevin is with me there.

Patch-set: 4
Reviewer: Paul Keith <11801@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>