audiopolicy: evaluate all outputs routing when call state changes
When call state changes, it is possible that some outputs are active in
which case their routing must be reevaluated.
This generalizes this behavior already implemented when a device is
connected/disconnected or a forced usage changed.

This is normally not needed as music is paused when a call is active
but there can be some race conditions where music resumes before the
audio mode is actually updated due to the asynchronous nature of audio
focus and audio mode management.

Bug: 78248310
Test: repro steps in bug. Audio smoke tests
Change-Id: If491eb879a90cf29025bedb2bc77210f052bcb6a