VT: Retain RTP sequence number between SFR stop and start
There was an issue that remote side dropped a video packets
when local camera has been switched.

It caused we are sending discontinuous (RTP) sequence number
when the camera is switched during video call is going on.

When we switch a camera, we just stopped a MediaRecorder
and then re-start with different camera-id.

RTP sequence number is gone when an ARTPWriter is stopped.
and Then ARTPWriter will take a random RTP sequence number from
the re-start of ARTPWriter.

This patch makes ARTPWriter can start from the RTP sequence number
that the stopped last time.

Bug: 121230209
Change-Id: I7dde2c89f2b378600a080b7a4f4de385f81e6ee1
Signed-off-by: Somraj Mani <som.mani@samsung.com>
Signed-off-by: Kim Sungyeon <sy85.kim@samsung.com>