Unmap shared memory when no longer used
CryptoHal in mediadrmserver was not releasing the
shared memory allocated remotely in the drm HAL
module until the HAL was closed, which could cause
shared memory allocation failures in situations
where multiple memory regions get mapped during
playback. This change releases the shared mapping
in the HAL when it is no longer needed.

bug:80104026

test:
 1. manual tests with logging to ensure
   shared memory is released
 2. gts media tests on walleye

Change-Id: Ie2f306bce6aec697ae584da3f55b3cf72edaa07b