audio polcy: fix explicit device routing in call
Make sure tracks with an explicit routing request are not systematically invalidated
when starting while another track on the same strategy is already active.

Also fix a problem where explicit routing is not taken into account by
getNewOutputDevice() when in call because the strategy with the
explicit routing is not evaluated.

Bug: 69680975
Bug: 79878501
Test: voice call regression, CTS for explicit audio routing, test case
for b/79878501

Change-Id: I99b83f8a3304549bf327c3c327f6671f777cfb4b