audiopolicy: apm: connectAudioSource: use Sw brigding according to hal support
Even if the implementation is not ready, this patch updates the conditions
of audio patch creation delegation to audio hal not only if HAL
is at least supporting 3.0 routing API but also has declared within
policy configuration file that a route between Device port is supported.

If the route is not declared in configuration file, Audio HAL is not able
to do the connexion, let AudioFlinger do the the Sw Bridging.

Test: Audio smoke tests.
Test: CTS tests for AudioRecord, AudioTrack and AudioEffect

Change-Id: I6ed11be04473be2edfcf16e922d1c6864184b6cd
Signed-off-by: François Gaffie <francois.gaffie@renault.com>