audioflinger: Add support for RecordTrack with no conversion
When piping encoded audio data via software patch, it needs
to be created without the buffer converter. In this case the audio
data is copied directly from ResamplerBufferProvider.

Added input flag AUDIO_INPUT_FLAG_DIRECT to indicate to PatchPanel
that the PatchRecord track needs to be opened in this mode.
This flag can be later added to minor Audio HAL update.

Bug: 63901775
Test: MSD scenario; verified that phone over USB headset still works
      on marlin / sailfish

Change-Id: If2745778d356769b143fb3c29910275ddef119ac